refactor(cli): remove legacy config module
Move policy-derived analysis tree helpers into a dedicated analysis module and
delete the old TOML config types. Simplify config mode resolution to policy
files only, stop writing Hipcheck.toml during setup, and remove the unused TOML
reader.

Work performance by GPT-5, reviewed by me.
